# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Meta API",
    "description": "示範範例生成文件"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "headers",
      "name": "authorization",
      "description": "請加上 API Token"
    }
  },
  "consumes": [],
  "produces": [],
  "paths": {
    "/api/posts": {
      "get": {
        "tags": [
          "Posts - 貼文"
        ],
        "description": "取得所有貼文",
        "parameters": [
          {
            "name": "timeSort",
            "in": "query"
          },
          {
            "name": "keyword",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "63306b16f2b86f0db81b4b3c"
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "63306ad1f2b86f0db81b4b28"
                          },
                          "name": {
                            "type": "string",
                            "example": "John"
                          },
                          "photo": {
                            "type": "string",
                            "example": "vvvv"
                          }
                        }
                      },
                      "tags": {
                        "type": "string",
                        "example": "忍者"
                      },
                      "type": {
                        "type": "string",
                        "example": "person"
                      },
                      "image": {
                        "type": "string",
                        "example": ""
                      },
                      "content": {
                        "type": "string",
                        "example": "我是漩渦鳴人"
                      },
                      "likes": {
                        "type": "number",
                        "example": 0
                      },
                      "comments": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Posts - 貼文"
        ],
        "description": "新增貼文",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "type": "object",
            "required": true,
            "description": "資料格式",
            "schema": {
              "$ref": "#/definitions/createdPostsSchema"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Some description...",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean",
                  "example": true
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "63306b16f2b86f0db81b4b3c"
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string",
                            "example": "63306ad1f2b86f0db81b4b28"
                          },
                          "name": {
                            "type": "string",
                            "example": "John"
                          },
                          "photo": {
                            "type": "string",
                            "example": "vvvv"
                          }
                        }
                      },
                      "tags": {
                        "type": "string",
                        "example": "忍者"
                      },
                      "type": {
                        "type": "string",
                        "example": "person"
                      },
                      "image": {
                        "type": "string",
                        "example": ""
                      },
                      "content": {
                        "type": "string",
                        "example": "我是漩渦鳴人"
                      },
                      "likes": {
                        "type": "number",
                        "example": 0
                      },
                      "comments": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Posts - 貼文"
        ],
        "description": "刪除所有貼文",
        "parameters": [],
        "responses": {}
      }
    },
...
  },
  "definitions": {
    "createdPostsSchema": {
      "type": "object",
      "properties": {
        "user": {
          "type": "string",
          "example": "63306ad1f2b86f0db81b4b28"
        },
        "tags": {
          "type": "string",
          "example": "忍者"
        },
        "type": {
          "type": "string",
          "example": "person"
        },
        "content": {
          "type": "string",
          "example": "我是漩渦鳴人"
        }
      },
      "required": [
        "user",
        "tags",
        "type",
        "content"
      ]
    }
  }
}